requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic>=2.9.0
sentence-transformers==2.7.0
faiss-cpu>=1.9.0
rank-bm25==0.2.2
google-genai>=1.0.0
numpy>=2.0.0
aiofiles==23.2.1
python-dotenv==1.0.1
torch>=2.6.0
transformers>=4.41.1
scikit-learn>=1.4.2
